@@ -17,7 +17,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3",
+        "php": "~7.0",
         "ext-zmq": "*",
 
         "neutron/process-manager": "2.0.x-dev@dev",
@@ -33,7 +33,7 @@
     ],
     "require-dev": {
         "monolog/monolog": "~1.0",
-        "phpunit/phpunit": "~3.7|^4.0|^5.0",
+        "phpunit/phpunit": "~6.0",
         "symfony/console": "~2.3",
         "symfony/finder": "~2.1"
     },
