@@ -2,15 +2,15 @@
 
 namespace Alchemy\Test\TaskManager;
 
-abstract class TestCase extends \PHPUnit_Framework_TestCase
+abstract class TestCase extends \PHPUnit\Framework\TestCase
 {
     public function createJobMock()
     {
-        return $this->getMock('Alchemy\TaskManager\Job\JobInterface');
+        return $this->createMock('Alchemy\TaskManager\Job\JobInterface');
     }
 
     public function createDataMock()
     {
-        return $this->getMock('Alchemy\TaskManager\Job\JobDataInterface');
+        return $this->createMock('Alchemy\TaskManager\Job\JobDataInterface');
     }
 }
