@@ -29,7 +29,7 @@ class MemoryLimitSubscriberTest extends SubscriberTestCase
         $job->expects($this->once())->method('stop');
         $job->expects($this->any())->method('isStarted')->will($this->returnValue(true));
 
-        $logger = $this->getMock('Psr\Log\LoggerInterface');
+        $logger = $this->createMock('Psr\Log\LoggerInterface');
         $logger->expects($this->once())->method('notice')->with('Max memory reached for romain (1 o.), stopping.');
 
         $subscriber = new MemoryLimitSubscriber(1, $logger);
