@@ -17,7 +17,7 @@ class StopSignalSubscriberTest extends SubscriberTestCase
         $job->expects($this->once())->method('stop');
         $job->expects($this->any())->method('isStarted')->will($this->returnValue(true));
 
-        $logger = $this->getMock('Psr\Log\LoggerInterface');
+        $logger = $this->createMock('Psr\Log\LoggerInterface');
         $logger->expects($this->once())->method('notice');
 
         $subscriber = new StopSignalSubscriber(SignalHandler::getInstance(), $logger);
